PatientVerification/requirements.txt
matplotlib==3.3.3
numpy==1.19.5
Pillow==8.1.0
scikit-learn==0.24.0
torch==1.7.1+cu110
torchvision==0.8.2+cu110
